third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-multicol/getclientrects-000-expected.txt

This is a testharness.js-based test.
[FAIL] horizontal-tb ltr
  assert_equals: expected 70 but got 69
[FAIL] vertical-lr ltr
  assert_equals: expected 120 but got 119
[FAIL] vertical-rl ltr
  assert_equals: expected 100 but got 99
[FAIL] horizontal-tb rtl
  assert_equals: expected 310 but got 332.21875
[FAIL] vertical-lr rtl
  assert_equals: expected 120 but got 119
[FAIL] vertical-rl rtl
  assert_equals: expected 100 but got 99
Harness: the test ran to completion.
